WIP: extract provider factory methods to own factory class
